Add initial common_test and PropEr testing
The speed test was moved from the vbisect.erl file to the common_test
vbisect_SUITE file. Cover analysis and testing are provided with HTML
reports for browsing the results.

Add is_vbisect/1,2 for testing opaque data values, as well as
data_version/1 to find out what version an opaque data structure is.

The current common_test speed test is 10x slower than compiling directly
on the shell command line. This is because of the test harness, not a
change to the implementation of vbisect. Future commits will address
this issue so that performance testing is more accurate.